fix(cors): 🐛 Fix cors issues
