+++ packages/service/src/server/index.js
@@ -2,6 +2,7 @@ import { ApolloServer } from 'apollo-server-express';
 import { config } from 'config';
 import http from 'http';
 import https from 'https';
+import { Cors } from 'types/cors';
 import { make as makeContext } from './context';
 import { formatError } from './formatError';
 import { make as makeHttpsOptions } from './httpsOptions';
@@ -25,7 +26,10 @@ const makeServer = ({ cache, context } = {}) =>
 const make = async ({ app, cache }) => {
   const apolloServer = makeServer({ cache });
 
-  apolloServer.applyMiddleware({ app, cors: false });
+  apolloServer.applyMiddleware({
+    app,
+    cors: Cors.OPTIONS,
+  });
 
   let server = null;
